refactor: move vtkjs to a submodule and clean-up trailing spaces
